# Speed up channel config operations 