--- a/internal/rule/cache.go
+++ b/internal/rule/cache.go
@@ -2,6 +2,7 @@ package rule
 
 import (
 	"hash/fnv"
+	"sync"
 	"sync/atomic"
 	"time"
 )
@@ -13,6 +14,7 @@ type cacheItem struct {
 }
 
 type cacheShard struct {
+	mu     sync.RWMutex
 	size   int
 	buffer []cacheItem
 	index  int32
@@ -45,6 +47,8 @@ func (c *rollingCache) shardForKey(key string) *cacheShard {
 
 func (c *rollingCache) Get(channel string) (channelOptionsResult, bool) {
 	shard := c.shardForKey(channel)
+	shard.mu.RLock()
+	defer shard.mu.RUnlock()
 	for _, item := range shard.buffer {
 		if item.channel == channel && time.Now().Before(time.Unix(0, item.expires)) {
 			return item.value, true
@@ -55,6 +59,8 @@ func (c *rollingCache) Get(channel string) (channelOptionsResult, bool) {
 
 func (c *rollingCache) Set(channel string, value channelOptionsResult, ttl time.Duration) {
 	shard := c.shardForKey(channel)
+	shard.mu.Lock()
+	defer shard.mu.Unlock()
 	index := int(atomic.AddInt32(&shard.index, 1) % int32(shard.size))
 	item := cacheItem{
 		channel: channel,