## Proposed changes

Using `atomic.Value` to avoid config locks, adding rolling cache with item expiration for retrieving `ChannelOptions`.

Before:

```
BenchmarkContainer_ChannelOptions-12    	 8343092	       140.7 ns/op	      32 B/op	       1 allocs/op
BenchmarkContainer_Config-12            	23106529	        50.98 ns/op	       0 B/op	       0 allocs/op
```

After:

```
BenchmarkContainer_ChannelOptions-12    	22248208	        46.87 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainer_Config-12            	54884310	        21.48 ns/op	       0 B/op	       0 allocs/op
```